requirements.txt
nats-py>=2.7.0

